+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/property-type-form.tsx
@@ -1,14 +1,14 @@
 import { getPropertyTypeById } from "@hashintel/hash-subgraph/src/stdlib/element/property-type";
 
-import { useBlockProtocolGetPropertyType } from "../../../../../../../../components/hooks/block-protocol-functions/ontology/use-block-protocol-get-property-type";
+import { useBlockProtocolGetPropertyType } from "../../../../../../../components/hooks/block-protocol-functions/ontology/use-block-protocol-get-property-type";
 import {
   generateInitialTypeUri,
   TypeForm,
   TypeFormProps,
   useGenerateTypeBaseUri,
-} from "../../shared/type-form";
+} from "../shared/type-form";
 import { ExpectedValueSelector } from "./property-type-form/expected-value-selector";
-import { PropertyTypeFormValues } from "./property-type-form-values";
+import { PropertyTypeFormValues } from "./shared/property-type-form-values";
 
 export const PropertyTypeForm = (
   props: TypeFormProps<PropertyTypeFormValues>,
