+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/property-type-form/expected-value-selector.tsx
@@ -14,9 +14,10 @@ import {
 } from "react";
 import { useController, useFormContext, useWatch } from "react-hook-form";
 
-import { AutocompleteDropdown } from "../../../../../../../shared/autocomplete-dropdown";
-import { StyledPlusCircleIcon } from "../../../../../../../shared/styled-plus-circle-icon";
-import { PropertyTypeFormValues } from "../property-type-form-values";
+import { AutocompleteDropdown } from "../../../../../../shared/autocomplete-dropdown";
+import { StyledPlusCircleIcon } from "../../../../../../shared/styled-plus-circle-icon";
+import { dataTypeOptions } from "../shared/data-type-options";
+import { PropertyTypeFormValues } from "../shared/property-type-form-values";
 import { CustomExpectedValueBuilder } from "./expected-value-selector/custom-expected-value-builder";
 import { ExpectedValueChip } from "./expected-value-selector/expected-value-chip";
 import {
@@ -24,7 +25,6 @@ import {
   useCustomExpectedValueBuilderContext,
 } from "./expected-value-selector/shared/custom-expected-value-builder-context";
 import { expectedValuesOptions } from "./expected-value-selector/shared/expected-values-options";
-import { dataTypeOptions } from "./shared/data-type-options";
 
 const ExpectedValueSelectorDropdown = ({ children, ...props }: PaperProps) => {
   const {
