+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/property-type-form/expected-value-selector/custom-expected-value-builder/array-expected-value-builder/array-min-max-items.tsx
@@ -18,7 +18,7 @@ import {
 } from "react";
 import { useController, useFormContext, useWatch } from "react-hook-form";
 
-import { PropertyTypeFormValues } from "../../../../property-type-form-values";
+import { PropertyTypeFormValues } from "../../../../shared/property-type-form-values";
 
 type ItemInputProps = { width: string } & InputProps;
