+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/property-type-form/expected-value-selector/expected-value-chip.tsx
@@ -10,7 +10,7 @@ import {
 import { Box, chipClasses, Collapse, Stack, Typography } from "@mui/material";
 import { useState } from "react";
 
-import { ArrayType } from "../../property-type-form-values";
+import { ArrayType } from "../../shared/expected-value-types";
 import { expectedValuesOptions } from "./shared/expected-values-options";
 
 interface ExpectedValueChipProps {
