+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/property-type-form/expected-value-selector/shared/delete-expected-value-modal.tsx
@@ -8,7 +8,7 @@ import { Box, Stack, Typography } from "@mui/material";
 import { bindDialog, PopupState } from "material-ui-popup-state/hooks";
 import { Fragment } from "react";
 
-import { Modal } from "../../../../../../../../../../../components/modals/modal";
+import { Modal } from "../../../../../../../../../../components/modals/modal";
 
 type CountItemProps = { label: string; count: number };
