+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/property-type-form/expected-value-selector/shared/expected-values-options.ts
@@ -2,11 +2,11 @@ import { faList, faListCheck } from "@fortawesome/free-solid-svg-icons";
 import { theme } from "@hashintel/hash-design-system";
 import { types } from "@hashintel/hash-shared/ontology-types";
 
-import { fa100 } from "../../../../../../../../../../../shared/icons/pro/fa-100";
-import { faCube } from "../../../../../../../../../../../shared/icons/pro/fa-cube";
-import { faCubes } from "../../../../../../../../../../../shared/icons/pro/fa-cubes";
-import { faSquareCheck } from "../../../../../../../../../../../shared/icons/pro/fa-square-check";
-import { faText } from "../../../../../../../../../../../shared/icons/pro/fa-text";
+import { fa100 } from "../../../../../../../../../../shared/icons/pro/fa-100";
+import { faCube } from "../../../../../../../../../../shared/icons/pro/fa-cube";
+import { faCubes } from "../../../../../../../../../../shared/icons/pro/fa-cubes";
+import { faSquareCheck } from "../../../../../../../../../../shared/icons/pro/fa-square-check";
+import { faText } from "../../../../../../../../../../shared/icons/pro/fa-text";
 
 const chipColors = {
   blue: {
