+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/definition-tab/shared/insert-type-row.tsx
@@ -10,8 +10,8 @@ import { Dispatch, Ref, SetStateAction, useRef } from "react";
 import {
   TypeSelector,
   TypeSelectorType,
-} from "../property-list-card/insert-property-row/type-selector";
-import { withHandler } from "../property-list-card/shared/with-handler";
+} from "./insert-property-row/type-selector";
+import { withHandler } from "./with-handler";
 
 export type InsertTypeRowProps<T extends TypeSelectorType> = {
   inputRef: Ref<HTMLInputElement | null>;
