+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/definition-tab/shared/type-form.tsx
@@ -48,7 +48,7 @@ import {
 } from "react-hook-form";
 
 import { Modal } from "../../../../../../../components/modals/modal";
-import { withHandler } from "../property-list-card/shared/with-handler";
+import { withHandler } from "./with-handler";
 import { QuestionIcon } from "./question-icon";
 
 type TypeFormSubmitProps = Omit<
