+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/shared/use-property-types-context-value.tsx
@@ -1,8 +1,8 @@
 import { getRoots } from "@hashintel/hash-subgraph/src/stdlib/roots";
 import { useCallback, useEffect, useMemo, useState } from "react";
 
-import { useBlockProtocolAggregatePropertyTypes } from "../../../../../components/hooks/block-protocol-functions/ontology/use-block-protocol-aggregate-property-types";
-import { PropertyTypesContextValues } from "./shared/property-types-context";
+import { useBlockProtocolAggregatePropertyTypes } from "../../../../../../components/hooks/block-protocol-functions/ontology/use-block-protocol-aggregate-property-types";
+import { PropertyTypesContextValues } from "./property-types-context";
 
 export const usePropertyTypesContextValue = () => {
   const [propertyTypes, setPropertyTypes] = useState<
